Ref: structure of connector methods repository and exceptions
