+++ src/api/exceptions.py
@@ -1,7 +1,5 @@
 from fastapi import HTTPException
 
-from src.core.relationaldb.repositories.exceptions import NoRecordFoundError  #noqa: F401
-
 
 class ServerInitError(HTTPException):
     def __init__(self, code:int = 500, message:str = ''):
