+++ src/core/relationaldb/exceptions.py
@@ -23,3 +23,13 @@ class InvalidConnectorModeError(HTTPException):
             status_code=400,
             detail=f"Unknown connector mode '{mode}'. Use 'sync' or 'async'."
         )
+
+
+class RecordUpdateNotAllowedError(HTTPException):
+    def __init__(self, code:int = 404, message:str = ''):
+        super().__init__(status_code=code, detail=f"Record can not be modified\n {message}")
+
+
+class NoRecordFoundError(HTTPException):
+    def __init__(self, code:int = 404, message:str = ''):
+        super().__init__(status_code=code, detail=f"Record not found\n {message}")
